[Bug #1548] Add support for externalParentTerms (skos:broadMatch) to term creation.
Also simplify term creation action logic (URL resolution).
