+++ src/action/AsyncTermActions.ts
@@ -4,16 +4,16 @@
  *
  * This file contains asynchronous actions related to term management in the frontend.
  */
-import VocabularyUtils from "../util/VocabularyUtils";
+import VocabularyUtils, {IRI} from "../util/VocabularyUtils";
 import ActionType from "./ActionType";
 import { GetStoreState, ThunkDispatch } from "../util/Types";
 import {
-  asyncActionFailure,
-  asyncActionRequest,
-  asyncActionSuccess,
-  publishMessage,
+    asyncActionFailure,
+    asyncActionRequest,
+    asyncActionSuccess,
+    publishMessage, publishNotification,
 } from "./SyncActions";
-import Ajax, { param } from "../util/Ajax";
+import Ajax, {content, param} from "../util/Ajax";
 import Term from "../model/Term";
 import { ErrorData } from "../model/ErrorInfo";
 import Constants from "../util/Constants";
@@ -22,6 +22,61 @@ import Message from "../model/Message";
 import MessageType from "../model/MessageType";
 import { getLocalized } from "../model/MultilingualString";
 import { getShortLocale } from "../util/IntlUtil";
+import Utils from "../util/Utils";
+import {AxiosResponse} from "axios";
+import * as SyncActions from "./SyncActions";
+
+export function createTerm(term: Term, vocabularyIri: IRI) {
+    const action = {
+        type: ActionType.CREATE_VOCABULARY_TERM,
+    };
+    return (dispatch: ThunkDispatch) => {
+        dispatch(asyncActionRequest(action));
+        const url = resolveTermCreationUrl(term, vocabularyIri);
+        const data = Object.assign(term.toJsonLd(), {
+            vocabulary: {
+                iri: vocabularyIri.namespace + vocabularyIri.fragment,
+            },
+        });
+        return Ajax.post(
+            url,
+            content(data)
+                .contentType(Constants.JSON_LD_MIME_TYPE)
+                .param("namespace", vocabularyIri.namespace)
+        )
+            .then((resp: AxiosResponse) => {
+                const asyncSuccessAction = asyncActionSuccess(action);
+                dispatch(asyncSuccessAction);
+                dispatch(
+                    SyncActions.publishMessage(
+                        new Message(
+                            { messageId: "vocabulary.term.created.message" },
+                            MessageType.SUCCESS
+                        )
+                    )
+                );
+                dispatch(publishNotification({ source: asyncSuccessAction }));
+                return resp.headers[Constants.Headers.LOCATION];
+            })
+            .catch((error: ErrorData) => {
+                dispatch(asyncActionFailure(action, error));
+                dispatch(
+                    SyncActions.publishMessage(new Message(error, MessageType.ERROR))
+                );
+                return undefined;
+            });
+    };
+}
+
+function resolveTermCreationUrl(term: Term, targetVocabularyIri: IRI) {
+    let url = `${Constants.API_PREFIX}/vocabularies/${targetVocabularyIri.fragment}/terms`;
+    const parents = Utils.sanitizeArray(term.parentTerms);
+    if (parents.length > 0) {
+        // Use one of the parents, it does not matter which one
+        url += `/${VocabularyUtils.create(parents[0].iri!).fragment}/subterms`;
+    }
+    return url;
+}
 
 export function setTermDefinitionSource(source: TermOccurrence, term: Term) {
   const termIri = VocabularyUtils.create(term.iri);
