+++ src/action/__tests__/AsyncActions.test.ts
@@ -1,9 +1,8 @@
-import configureMockStore, { MockStoreEnhanced } from "redux-mock-store";
+import configureMockStore, {MockStoreEnhanced} from "redux-mock-store";
 import {
   createFileInDocument,
   createProperty,
   createResource,
-  createTerm,
   createVocabulary,
   executeFileTextAnalysis,
   exportFileContent,
@@ -12,12 +11,15 @@ import {
   getLabel,
   getProperties,
   hasFileContent,
+  loadAllTerms,
+  loadConfiguration,
   loadFileContent,
   loadHistory,
   loadImportedVocabularies,
   loadLastEditedAssets,
   loadLatestTextAnalysisRecord,
   loadMyAssets,
+  loadNews,
   loadResource,
   loadResources,
   loadResourceTermAssignmentsInfo,
@@ -26,7 +28,6 @@ import {
   loadTerms,
   loadTypes,
   loadUnusedTermsForVocabulary,
-  validateVocabulary,
   loadVocabularies,
   loadVocabulary,
   loadVocabularyContentChanges,
@@ -39,58 +40,42 @@ import {
   updateTerm,
   updateVocabulary,
   uploadFileContent,
-  loadNews,
-  loadConfiguration,
-  loadAllTerms,
+  validateVocabulary,
 } from "../AsyncActions";
 import Constants from "../../util/Constants";
-import Ajax, { param } from "../../util/Ajax";
+import Ajax, {param} from "../../util/Ajax";
 import thunk from "redux-thunk";
-import { Action } from "redux";
+import {Action} from "redux";
 import Routing from "../../util/Routing";
-import Vocabulary, {
-  CONTEXT as VOCABULARY_CONTEXT,
-} from "../../model/Vocabulary";
-import Vocabulary2, { IRI } from "../../util/VocabularyUtils";
-import VocabularyUtils from "../../util/VocabularyUtils";
+import Vocabulary, {CONTEXT as VOCABULARY_CONTEXT,} from "../../model/Vocabulary";
+import Vocabulary2 from "../../util/VocabularyUtils";
+import VocabularyUtils, {IRI} from "../../util/VocabularyUtils";
 import Routes from "../../util/Routes";
-import ActionType, {
-  AsyncAction,
-  AsyncActionSuccess,
-  MessageAction,
-} from "../ActionType";
-import Term, { CONTEXT as TERM_CONTEXT } from "../../model/Term";
+import ActionType, {AsyncAction, AsyncActionSuccess, MessageAction,} from "../ActionType";
+import Term from "../../model/Term";
 import Generator from "../../__tests__/environment/Generator";
-import { ThunkDispatch } from "../../util/Types";
+import {ThunkDispatch} from "../../util/Types";
 import FetchOptionsFunction from "../../model/Functions";
-import RdfsResource, {
-  CONTEXT as RDFS_RESOURCE_CONTEXT,
-} from "../../model/RdfsResource";
+import RdfsResource, {CONTEXT as RDFS_RESOURCE_CONTEXT,} from "../../model/RdfsResource";
 import TermItState from "../../model/TermItState";
-import Resource, { CONTEXT as RESOURCE_CONTEXT } from "../../model/Resource";
+import Resource, {CONTEXT as RESOURCE_CONTEXT} from "../../model/Resource";
 import Utils from "../../util/Utils";
 import AsyncActionStatus from "../AsyncActionStatus";
 import ExportType from "../../util/ExportType";
 import fileContent from "../../rest-mock/file";
 import TermItFile from "../../model/File";
 import MessageType from "../../model/MessageType";
-import {
-  CONTEXT as TA_RECORD_CONTEXT,
-  TextAnalysisRecord,
-} from "../../model/TextAnalysisRecord";
+import {CONTEXT as TA_RECORD_CONTEXT, TextAnalysisRecord,} from "../../model/TextAnalysisRecord";
 import {
   CONTEXT as RESOURCE_TERM_ASSIGNMENT_CONTEXT,
   ResourceTermAssignments,
 } from "../../model/ResourceTermAssignments";
-import {
-  CONTEXT as TERM_ASSIGNMENTS_CONTEXT,
-  TermAssignments,
-} from "../../model/TermAssignments";
-import { verifyExpectedAssets } from "../../__tests__/environment/TestUtil";
+import {CONTEXT as TERM_ASSIGNMENTS_CONTEXT, TermAssignments,} from "../../model/TermAssignments";
+import {verifyExpectedAssets} from "../../__tests__/environment/TestUtil";
 import ChangeRecord from "../../model/changetracking/ChangeRecord";
 import RecentlyModifiedAsset from "../../model/RecentlyModifiedAsset";
 import NotificationType from "../../model/NotificationType";
-import { langString } from "../../model/MultilingualString";
+import {langString} from "../../model/MultilingualString";
 import ValidationResult from "../../model/ValidationResult";
 import UserRole from "../../model/UserRole";
 
@@ -421,154 +406,6 @@ describe("Async actions", () => {
         )
       ).then(() => {
         expect(Ajax.get).not.toHaveBeenCalled();
-      });
-    });
-  });
-
-  describe("create term", () => {
-    const vocabularyIri = VocabularyUtils.create(
-      "http://onto.fel.cvut.cz/ontologies/termit/vocabulary/test-vocabulary"
-    );
-
-    it("create top level term in vocabulary context and send it over the network", () => {
-      const term = new Term({
-        label: langString("Test term 1"),
-        iri: vocabularyIri.toString() + "term/test-term-1",
-      });
-      const mock = jest.fn().mockImplementation(() => Promise.resolve());
-      Ajax.post = mock;
-      return Promise.resolve(
-        (store.dispatch as ThunkDispatch)(createTerm(term, vocabularyIri))
-      ).then(() => {
-        expect(Ajax.post).toHaveBeenCalled();
-        expect(mock.mock.calls[0][0]).toEqual(
-          Constants.API_PREFIX +
-            "/vocabularies/" +
-            vocabularyIri.fragment +
-            "/terms"
-        );
-        const config = mock.mock.calls[0][1];
-        expect(config.getContentType()).toEqual(Constants.JSON_LD_MIME_TYPE);
-        const data = config.getContent();
-        expect(data["@context"]).toBeDefined();
-        expect(data["@context"]).toEqual(TERM_CONTEXT);
-      });
-    });
-
-    it("create child term in vocabulary context and send it over the network", () => {
-      const parentFragment = "test-term-1";
-      const parentTerm = new Term({
-        label: langString("Test term 1"),
-        iri: vocabularyIri.toString() + "term/" + parentFragment,
-        vocabulary: { iri: vocabularyIri.toString() },
-      });
-      const childTerm = new Term({
-        label: langString("Test term 2"),
-        iri: vocabularyIri.toString() + "term/test-term-2",
-        parentTerms: [parentTerm],
-      });
-      const mock = jest.fn().mockImplementation(() => Promise.resolve());
-      Ajax.post = mock;
-      return Promise.resolve(
-        (store.dispatch as ThunkDispatch)(createTerm(childTerm, vocabularyIri))
-      ).then(() => {
-        expect(Ajax.post).toHaveBeenCalled();
-        expect(mock.mock.calls[0][0]).toEqual(
-          Constants.API_PREFIX +
-            "/vocabularies/" +
-            vocabularyIri.fragment +
-            "/terms/" +
-            parentFragment +
-            "/subterms"
-        );
-        const config = mock.mock.calls[0][1];
-        expect(config.getContentType()).toEqual(Constants.JSON_LD_MIME_TYPE);
-        const data = config.getContent();
-        expect(data["@context"]).toBeDefined();
-        expect(data["@context"]).toEqual(TERM_CONTEXT);
-      });
-    });
-
-    it("publishes notification on successful creation", () => {
-      const term = new Term({
-        label: langString("Test term 1"),
-        iri: vocabularyIri.toString() + "term/test-term-1",
-      });
-      Ajax.post = jest.fn().mockImplementation(() =>
-        Promise.resolve({
-          headers: {
-            location: "http://test",
-          },
-        })
-      );
-      return Promise.resolve(
-        (store.dispatch as ThunkDispatch)(createTerm(term, vocabularyIri))
-      ).then(() => {
-        const actions = store.getActions();
-        const action = actions[actions.length - 1];
-        expect(action.type).toEqual(ActionType.PUBLISH_NOTIFICATION);
-        expect(action.notification.source.type).toEqual(
-          ActionType.CREATE_VOCABULARY_TERM
-        );
-        expect(action.notification.source.status).toEqual(
-          AsyncActionStatus.SUCCESS
-        );
-      });
-    });
-
-    it("provides vocabulary namespace in a request parameter", () => {
-      const term = new Term({
-        label: langString("Test term 1"),
-        iri: vocabularyIri.toString() + "term/test-term-1",
-      });
-      const mock = jest.fn().mockImplementation(() => Promise.resolve());
-      Ajax.post = mock;
-      return Promise.resolve(
-        (store.dispatch as ThunkDispatch)(createTerm(term, vocabularyIri))
-      ).then(() => {
-        const config = mock.mock.calls[0][1];
-        expect(config.getParams().namespace).toEqual(vocabularyIri.namespace);
-      });
-    });
-
-    it("uses parent vocabulary IRI for request URL when new child term is in different vocabulary than parent term", () => {
-      const parentVocabularyIri =
-        "http://onto.fel.cvut.cz/ontologies/termit/vocabulary/parent-vocabulary";
-      const parentTerm = new Term({
-        iri: parentVocabularyIri + "/terms/parent-term",
-        label: langString("Parent term"),
-        vocabulary: { iri: parentVocabularyIri },
-      });
-      const childTerm = new Term({
-        iri: vocabularyIri.toString() + "/terms/test-term",
-        label: langString("Test term"),
-        parentTerms: [parentTerm],
-      });
-      Ajax.post = jest.fn().mockImplementation(() => Promise.resolve());
-      return Promise.resolve(
-        (store.dispatch as ThunkDispatch)(createTerm(childTerm, vocabularyIri))
-      ).then(() => {
-        const url = (Ajax.post as jest.Mock).mock.calls[0][0];
-        const config = (Ajax.post as jest.Mock).mock.calls[0][1];
-        const parentIri = VocabularyUtils.create(parentVocabularyIri);
-        expect(url).toContain(parentIri.fragment);
-        expect(config.getParams().namespace).toEqual(parentIri.namespace);
-      });
-    });
-
-    it("sets term vocabulary before sending it to server", () => {
-      const term = new Term({
-        label: langString("Test term 1"),
-        iri: vocabularyIri.toString() + "term/test-term-1",
-      });
-      Ajax.post = jest.fn().mockImplementation(() => Promise.resolve());
-      return Promise.resolve(
-        (store.dispatch as ThunkDispatch)(createTerm(term, vocabularyIri))
-      ).then(() => {
-        const config = (Ajax.post as jest.Mock).mock.calls[0][1];
-        const data = config.getContent();
-        expect(data.vocabulary).toBeDefined();
-        expect(data.vocabulary.iri).toEqual(vocabularyIri.toString());
       });
     });
   });
