+++ src/component/term/CreateTerm.tsx
@@ -3,7 +3,8 @@ import Term from "../../model/Term";
 import VocabularyUtils, { IRI } from "../../util/VocabularyUtils";
 import { connect } from "react-redux";
 import { ThunkDispatch } from "../../util/Types";
-import { createTerm, loadVocabulary } from "../../action/AsyncActions";
+import {createTerm} from "../../action/AsyncTermActions";
+import { loadVocabulary } from "../../action/AsyncActions";
 import TermMetadataCreate from "./TermMetadataCreate";
 import Routing from "../../util/Routing";
 import Routes from "../../util/Routes";
