+++ src/component/term/TermMetadataCreateForm.tsx
@@ -21,6 +21,7 @@ import { checkLabelUniqueness } from "./TermValidationUtils";
 import ShowAdvancedAssetFields from "../asset/ShowAdvancedAssetFields";
 import { loadIdentifier } from "../asset/AbstractCreateAsset";
 import MultilingualIcon from "../misc/MultilingualIcon";
+import {TermMetadataEdit} from "./TermMetadataEdit";
 
 interface TermMetadataCreateFormProps extends HasI18n {
   onChange: (change: object, callback?: () => void) => void;
@@ -131,8 +132,9 @@ export class TermMetadataCreateForm extends React.Component<
     this.props.onChange({ types });
   };
 
-  public onParentSelect = (parentTerms: Term[]) => {
-    this.props.onChange({ parentTerms });
+  public onParentSelect = (value: Term[]) => {
+    const split = TermMetadataEdit.splitTermsInSameAndDifferentVocabularies(value, this.props.vocabularyIri);
+    this.props.onChange({ parentTerms: split.sameVocabulary, externalParentTerms: split.differentVocabulary });
   };
 
   public render() {
@@ -213,7 +215,7 @@ export class TermMetadataCreateForm extends React.Component<
             <ParentTermSelector
               id="create-term-parent"
               onChange={this.onParentSelect}
-              parentTerms={termData.parentTerms}
+              parentTerms={[...Utils.sanitizeArray(termData.parentTerms), ...Utils.sanitizeArray(termData.externalParentTerms)]}
               vocabularyIri={this.props.vocabularyIri}
             />
           </Col>
